```
================================================================================
Valkey JSON 1.0.2 - Released Mon 8 September 2025
================================================================================
* Support building valkey-json on Clang/Apple (#69)
* Workdflow to automatically trigger valkey-bundle release (#64)
* Allow testing on external container server (#70)
* Implement SharedAPI interface that mimics JSON.GET (#71)
```

+++ tst/unit/hashtable_test.cc
@@ -73,7 +73,7 @@ static JValue makeArray(size_t sz, size_t offset = 0) {
     JValue j;
     j.SetArray();
     for (size_t i = 0; i < sz; ++i) {
-        j.PushBack(JValue(i + offset), allocator);
+        j.PushBack(JValue(static_cast<uint64_t>(i + offset)), allocator);
     }
     return j;
 }
